test: UserRepository 관련 테스트 생성 및 테스트에 맞게 도메인 수정

+++ src/main/java/me/siyoon/noticeboard/repository/NoticeContentRepository.java
@@ -3,5 +3,5 @@ package me.siyoon.noticeboard.repository;
 import me.siyoon.noticeboard.domain.NoticeContent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NoticeContentRepository extends JpaRepository<NoticeContent, Integer> {
+public interface NoticeContentRepository extends JpaRepository<NoticeContent, Long> {
 }
